refactor(images): read images from cloudflare
