+++ src/components/medical-dashboard/treatmentCard/TreatmentNotFound/TreatmentNotFound.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Card } from '../../../common/Card/Card';
-import notFoundImg from 'assets/images/nothing-found.png';
+import notFoundImg from 'assets/images/nothing-found.webp';
 import { useTranslation } from 'react-i18next';
 import * as S from './TreatmentNotFound.styles';
